requirements.txt
certifi==2019.9.11
chardet==3.0.4
numpy==1.17.4
pandas==0.25.3
pdfminer.six==20191110
pycryptodome==3.9.3
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sortedcontainers==2.1.0
